거점 로직 작성, 거점 패배 로직, 그리드랑 포탑 배치 -> 상점에서 거점강화

+++ Assets/_Project/_Scripts/Tower/Turret.cs
@@ -8,7 +8,7 @@ namespace Defense
         [Header("Turret Base Settings")]
         public float attackRange = 5f;
         public float fireRate = 1f;
-        public float attackDamage = 10f;
+        public int attackDamage = 10;
         public LayerMask enemyLayer;
 
         [Header("Debug / Test")]
